requirements.txt
python-telegram-bot==22.1
python-dotenv==1.0.0

